[issue-1152] Improve performance Browser
